Implement standalone compilation for AllocatorLib
This commit modifies the CMakeLists.txt file to allow the AllocatorLib to be either built standalone or as a part of a larger project. It adds conditions where if the source directory is the same as the current directory, it includes all c and h files within the 'source' and 'executable' directories to be compiled into the executable 'AllocatorLib'.
Also, '.gitignore' is updated to ignore the 'cmake-build-debug' directory and an empty 'main.c' file is added. This setup provides a flexible build configuration for the AllocatorLib.