Add functionality for memory descriptor generation
The main.c file has been updated with functions to generate Memory Descriptors and Kernel Parameters. These functions will randomly generate parameters for Memory Descriptors and return a pointer to a KERNEL_PARAMETERS structure. This is useful for testing memory handling and allocation. Previous implementation was just printing a 'Hello World!' string which has been replaced with this functionality plus a more detailed output of the generated data.